+++ tests/assets_manager_test.cc
@@ -150,3 +150,24 @@ TEST(AssetsManager, any) {
   ASSERT_EQ(strncmp((const char*)(r->data), "abc\n", 4), 0);
 #endif /*WITH_FS_RES*/
 }
+
+static asset_info_t* custom_load_asset(assets_manager_t* am, asset_type_t type, const char* name) {
+  asset_info_t* info = asset_info_create(ASSET_TYPE_DATA, 0, "test.any", 5);
+  memcpy(info->data, "abcd", 5);
+
+  return info;
+}
+
+TEST(AssetsManager, custom_load_asset) {
+  const asset_info_t* r = NULL;
+  assets_manager_t* rm = assets_manager();
+
+  assets_manager_set_custom_load_asset(rm, custom_load_asset, NULL); 
+  r = assets_manager_ref(rm, ASSET_TYPE_DATA, "test.any");
+
+  ASSERT_EQ(r != NULL, true);
+  ASSERT_STREQ((const char*)(r->data), "abcd");
+  assets_manager_set_custom_load_asset(rm, NULL, NULL); 
+
+  assets_manager_unref(rm, r);
+}